FIX: move from pnpm to npm due to typegen incompatibility and missing peer deps
