+++ sanity-typegen.json
@@ -1,5 +1,5 @@
 {
-    "path": "./src/**/*.{ts,tsx,js,jsx}", 
+    "path": "./sanity/**/*.{ts,tsx,js,jsx}", 
     "schema": "./sanity/types/schema.json",
     "generates": "./sanity/types/sanity.types.ts",
     "overloadClientMethods": true
