+++ sanity.config.ts
@@ -10,7 +10,7 @@ import {structureTool} from 'sanity/structure'
 
 // Go to https://www.sanity.io/docs/api-versioning to learn how API versioning works
 import {apiVersion, dataset, projectId} from './sanity/env'
-import {schema} from './sanity/schemaTypes'
+import {schema} from './sanity/schema'
 import {structure} from './sanity/structure'
 
 interface Group {
