+++ sanity/lib/groq.ts
@@ -1,17 +1,8 @@
 import { sanityFetch } from "@/app/studio/live";
 import { client } from './client';
-import { defineQuery } from "next-sanity";
 import type { Post } from '../types/sanity.types'
+import { POSTS_QUERY } from "./sanity.queries";
 
-const POSTS_QUERY = defineQuery(`
- *[_type == "post" && defined(slug.current)]{
-        _id,
-        name,
-        "slug": slug.current,
-        date
-    } | order(date desc)
-    `)
-
-export function getAllPosts(): Promise<Post[]> {
+export function getAllPosts(){
     return client.fetch(POSTS_QUERY)
 }
